backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
pydantic-settings
httpx